feat: replace deprecated apt-key with a repository signed-by a PGP key
This is a proposal to update Tailscale installation process to follow
current guidelines for managing third party repositories for
Debian-based distributions.
https://wiki.debian.org/DebianRepository/UseThirdParty

Current Tailscale docs suggest to store the GPG key in
/usr/share/keyrings staring from Debian 11 "bullseye" and Ubuntu 20.04,
but proposed changes should work for distributions based on Debian
version 9 "stretch" and higher.
https://tailscale.com/kb/1038/install-debian-bullseye/

The usage of apt-key tool to manage keys is deprecated in Ubuntu 22.04
and will be removed in the upcoming release. Use of apt-key tool
to manage keys results in a deprecation warning printed in the output
of every interaction with apt command. The proposed fix follows an
example provided in the Ansible documentation for the apt-key module
https://docs.ansible.com/ansible/latest/collections/ansible/builtin/apt_key_module.html#notes