feat: Allow ignoring auth key change
This adds an option to not include the authkey in state file. It might
be useful for short-lived keys that are generated for each run and would
otherwise lead to a "change" on every run.

The functionality is driven by `auth_key_in_state` and it defaults to
True, so by default the current behavior of having the key in state is
preserved.

While working on this, I decided to simplify the state handling a bit:
* Dropped the state fetch and hash comparison functionality. Behind the
  scenes, Ansible will do this anyways (to figure out if file needs to
  be updated) so we can just rely on the "changed" attribute of the
  state file instead.
* Set the default of `tailscale_authkey` and `tailscale_args` to empty
  string rather than null. This still evaluates to False in boolean
  context and it removes the need to use `| default()` frequently in
  contexts where the values might be optional.
* Removed couple `| bool` filters in conditional context where bool is
  implied anyways by Ansible.