feat: Use local Headscale instance for molecule testing (#319)
* Use local Headscale instance for molecule testing

This adds Headscale container setup to Molecule scenarios to be used for
testing when `TAILSCALE_CI_KEY` is not set to enable easier local
testing.

If the env variable is set, Molecule tests will revert back to the
previous behavior and use Tailscale key for testing.

* Drive the Headscale usage with env variable

As per feedback in [PR][0] this change actually requires `USE_HEADSCALE`
variable to be set before attempting to use local Headscale instance.
This is to avoid silently succeeding the CI against Headscale and
releasing code that wasn't tested with Tailscale.

Also added was documentation describing how to test against local
Headscale instance.

[0]: https://github.com/artis3n/ansible-role-tailscale/pull/319

* Avoid linting headscale config by ansible-lint

Ansible lint assumed this is a playbook and reported failure. This
should tell ansible-lint to avoid linting this file.

* Avoid linting tasks file

ansible-lint attempts to lint init_tailscale_vars.yml as playbook, but
this is actually tasks file, so we need to tell linter to skip it.

* Add community.docker to dev requirements

This is used to prepare Headscale container before converge runs in
Molecule, so we should add it to the requirements.yml

* Add a link to headscale configuration doc

We're only using subset of settings and leave the rest on default. For
better understanding of what's set in the configuration, let's add a
link to the example file which contains explanation for each setting.

* Also run test using Headscale

Add new job to test the role against Headscale instance.

* Move headscale workflow job

Move the job from `molecule.yml` to the end of `pull_request_target.yml`
to only run this test for one OS rather than the entire matrix of
configurations.

* Set the job runner

Co-authored-by: Ari Kalfus <dev@artis3nal.com>

---------

Co-authored-by: Ari Kalfus <dev@artis3nal.com>