Make tailscale status parsing less fragile
Tailscale has an option to print out json instead. This feels like much
less fragile way of checking status as we're not reliant on some
heuristics and string matching. Instead we can look at very specific
field.

This also opens up some option in the future to look at more detailed
aspects of the state and perhaps react to it in more nuanced way.