Workflow and testing improvements (#328)
* Add static mapping for headscale to hosts file

Tailscale does attempt to manage `resolv.conf` which is somewhat in
conflict with Docker attempting to provice container name resolution.

This extra steps add static hosts override for headscale container so
that Tailscale managing DNS resolution does not mess up the connectivity
to Headscale container.

* Make tailscale status parsing less fragile

Tailscale has an option to print out json instead. This feels like much
less fragile way of checking status as we're not reliant on some
heuristics and string matching. Instead we can look at very specific
field.

This also opens up some option in the future to look at more detailed
aspects of the state and perhaps react to it in more nuanced way.

* Fix issues raised by lint

* Fix the tailscale up run logic

As per [comment here][0], the logic should be that either state changed
or client is not online.

[0]: https://github.com/artis3n/ansible-role-tailscale/pull/328/files#r1186696781

* Add explanation for the need of unsafe_writes

* ci: run headscale through several scenarios

---------

Co-authored-by: Artis3n <dev@artis3nal.com>